Fix scaffold (#265)
Remove SingleChildScrollView as inner screens don't render when it is set on the scaffold.
This also means that the login page would not scroll on a small screen but this should ideally be
fixed across the board with a consistent strategy. Requires some research on how to do this in flutter.